some more reorg
